pyproject.toml
[project]
name = "phoenix_dashboard"
version = "0.1.0"
description = "Silverback recorder and bot controller client"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "jinja2>=3.1.4",
  "sse-starlette>=2.1.3",
  "python-multipart",
  "web3>=6.0.0",
  "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
  "ruff",
  "black",
]

[tool.briefcase]
project_name = "Phoenix Dashboard"
bundle = "com.agworkywork.phoenix_dashboard"
version = "0.1.0"
url = "https://github.com/agBabbbyCakes/phoenix"
license = { text = "MIT" }
author = "Alex Gonzalez"
author_email = ""

[tool.briefcase.app.phoenix_dashboard]
formal_name = "Phoenix Dashboard"
description = "Silverback recorder and bot controller client"
sources = [
  "app",
  "src",
  "phoenix_dashboard",
]
requires = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.6",
  "jinja2>=3.1.4",
  "sse-starlette>=2.1.3",
  "python-multipart",
  "web3>=6.0.0",
  "python-dotenv>=1.0.0",
]
icon = "resources/phoenix"
document_types = []
resources = [
  "static",
  "templates",
]

[tool.briefcase.app.phoenix_dashboard.macOS]
requires = ">=10.14"

[tool.briefcase.app.phoenix_dashboard.windows]
requires = ">=10"

[tool.briefcase.app.phoenix_dashboard.linux]
requires = ">=18.04"
